simple_calculator_2.py
Someone recommended that I create a simple calculator in Python to become more comfortable with numbers, variables, functions, if statements, and boolean values. This was my second attempt at it. While the first attempt provided a functional calculator, the code was too repetitive and difficult to read. This calculator will allow the user to add, subtract, multiply, divide, and take exponents.