die_roll.py
This simple program uses the random module for Python to create rangebound random number generations, simulating a die roll. The user is able to select the number of sides the die has, as well as whether they would prefer the result be printed or returned by the function die_roll().