Fibonacci Tester
This program combines basic Python methods with number theory relating to the Fibonacci sequence and perfect squares to test whether or not the number entered as the 'number' variable is a part of the Fibonacci sequence. I am going to try to add a feature that gives the program the ability to tell whether or not the number is a prime number as well.